LayoutTests/platform/mac/imported/w3c/web-platform-tests/webxr/xrWebGLLayer_framebuffer_scale.https-expected.txt


FAIL Ensure framebuffer scaling works as expected. - webgl assert_approx_equals: expected NaN +/- 0.001 but got 1
FAIL Ensure framebuffer scaling works as expected. - webgl2 assert_approx_equals: expected NaN +/- 0.001 but got 1
